Добавление фильтров Доделали Journaling Доделали Controller
