+++ MyPetsForm.cs
@@ -56,64 +56,11 @@ namespace App_project
         {
             string connection = DataBase.PetDBConnectionString;
             DataBase.LinkDataBase();
-            
-            listView1.GridLines = false;
-            listView1.View = View.Details;
 
+            ListView list = listView1;
 
-            SqlConnection connection2 = DataBase.LinkDataBase();
+            Controller.LoadPetList(list);
 
-            string sql = "Select NickName, Category, Breed, PassportNumber from [PetDataBase].[dbo].[PetData] WHERE [IDUser] = '" + IDUser_key.global_IDUser + "'";
-            SqlConnection cnn = new SqlConnection(connection);
-            cnn.Open();
-            SqlCommand cmd = new SqlCommand(sql, cnn);
-            SqlDataReader Reader = cmd.ExecuteReader();
-
-            SqlCommand cmdPhoto = new SqlCommand("SELECT [Photo] FROM [PetDataBase].[dbo].[PetData] WHERE [IDUser] = '" + IDUser_key.global_IDUser + "'", connection2);
-            SqlDataAdapter dataAdapter = new SqlDataAdapter(cmdPhoto);
-            DataSet dataSet = new DataSet();
-            dataAdapter.Fill(dataSet);
-
-            listView1.Items.Clear();
-
-            ImageList imagelist = new ImageList();
-            imagelist.ImageSize = new Size(50, 50);
-            //if (dataSet.Tables[0].Rows.Count == 1)
-            //{
-
-            for (int i = 0; i < dataSet.Tables[0].Rows.Count; i++)
-            {
-                Byte[] data = new Byte[0];
-                data = (Byte[])(dataSet.Tables[0].Rows[i]["Photo"]);
-                MemoryStream mem = new MemoryStream(data);
-
-                imagelist.Images.Add(Image.FromStream(mem));
-            }
-
-            //}
-            listView1.SmallImageList = imagelist;
-            var nick = new List<string>();
-            var category = new List<string>();
-            var breed = new List<string>();
-            var passportn = new List<string>();
-
-            while (Reader.Read())
-            {
-                nick.Add(Reader.GetString(0));
-                category.Add(Reader.GetString(1));
-                breed.Add(Reader.GetString(2));
-                passportn.Add(Convert.ToString(Reader.GetInt32(3)));
-            }
-
-            for (int i = 0; i < category.Count; i++)
-            {
-                ListViewItem lst = new ListViewItem(new string[] { "", nick[i], category[i], breed[i], passportn[i] });
-                lst.ImageIndex = i;
-                listView1.Items.Add(lst);
-            }
-
-            Reader.Close();
-            cnn.Close();
         }
 
 
@@ -152,47 +99,53 @@ namespace App_project
         {
             string fileName = "C:\\Users\\kirillzagul\\Downloads\\ Список_карточек.xlsx";
 
-            try
-            {
-                var excel = new Excel.Application();
+            ListView list = listView1;
 
-                var workBooks = excel.Workbooks;
-                var workBook = workBooks.Add();
-                var workSheet = (Excel.Worksheet)excel.ActiveSheet;
-                workSheet.Columns[1].ColumnWidth = 30;
-                workSheet.Columns[2].ColumnWidth = 30;
-                workSheet.Columns[3].ColumnWidth = 30;
-                workSheet.Columns[4].ColumnWidth = 30;
-                var excelcells = workSheet.get_Range("A1", "D1");
-                excelcells.Borders.ColorIndex = 3;
-                //workSheet.Cells.Borders[].Weight = Excel.XlBorderWeight.xlThin;
-                //workSheet.Cells[1, "A"].Border.Weight = Excel.XlBorderWeight.xlThin;
-                //workSheet.Cells[1, "A"].Border.LineStyle =Excel.XlLineStyle.xlContinuous;
-                //workSheet.Cells[1, "A"].Border.ColorIndex = 0;
-                workSheet.Cells.Style.HorizontalAlignment = Excel.XlHAlign.xlHAlignCenter;
-                workSheet.Cells[1, "A"] = "Кличка";
-                workSheet.Cells[1, "B"] = "Вид";
-                workSheet.Cells[1, "C"] = "Порода";
-                workSheet.Cells[1, "D"] = "Номер паспорта";
-                for (int i = 0; i < listView1.Items.Count; i++)
-                {
-                    workSheet.Cells[i + 2, "A"] = listView1.Items[i].SubItems[1].Text;
-                    workSheet.Cells[i + 2, "B"] = listView1.Items[i].SubItems[2].Text;
-                    workSheet.Cells[i + 2, "C"] = listView1.Items[i].SubItems[3].Text;
-                    workSheet.Cells[i + 2, "D"] = listView1.Items[i].SubItems[4].Text;
-                }
+            Controller.ExportPetRegisterExcel(list);
 
-                //workSheet.Cells[6, "A"] = "Фото";
-                //workSheet.Cells[6, "B"] = pictureBox1.ImageLocation;
-                workBook.SaveAs(fileName);
-                workBook.Close();
+            //try
+            //{
+            //    var excel = new Excel.Application();
 
-                MessageBox.Show("Файл " + Path.GetFileName(fileName) + " записан успешно!");
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show("Ошибка: " + ex.ToString());
-            }
+            //    var workBooks = excel.Workbooks;
+            //    var workBook = workBooks.Add();
+            //    var workSheet = (Excel.Worksheet)excel.ActiveSheet;
+            //    workSheet.Columns[1].ColumnWidth = 30;
+            //    workSheet.Columns[2].ColumnWidth = 30;
+            //    workSheet.Columns[3].ColumnWidth = 30;
+            //    workSheet.Columns[4].ColumnWidth = 30;
+            //    var excelcells = workSheet.get_Range("A1", "D1");
+            //    excelcells.Borders.ColorIndex = 3;
+            //    //workSheet.Cells.Borders[].Weight = Excel.XlBorderWeight.xlThin;
+            //    //workSheet.Cells[1, "A"].Border.Weight = Excel.XlBorderWeight.xlThin;
+            //    //workSheet.Cells[1, "A"].Border.LineStyle =Excel.XlLineStyle.xlContinuous;
+            //    //workSheet.Cells[1, "A"].Border.ColorIndex = 0;
+            //    workSheet.Cells.Style.HorizontalAlignment = Excel.XlHAlign.xlHAlignCenter;
+            //    workSheet.Cells[1, "A"] = "Кличка";
+            //    workSheet.Cells[1, "B"] = "Вид";
+            //    workSheet.Cells[1, "C"] = "Порода";
+            //    workSheet.Cells[1, "D"] = "Номер паспорта";
+            //    for (int i = 0; i < list.Items.Count; i++)
+            //    {
+            //        workSheet.Cells[i + 2, "A"] = list.Items[i].SubItems[1].Text;
+            //        workSheet.Cells[i + 2, "B"] = list.Items[i].SubItems[2].Text;
+            //        workSheet.Cells[i + 2, "C"] = list.Items[i].SubItems[3].Text;
+            //        workSheet.Cells[i + 2, "D"] = list.Items[i].SubItems[4].Text;
+            //    }
+
+            //    //workSheet.Cells[6, "A"] = "Фото";
+            //    //workSheet.Cells[6, "B"] = pictureBox1.ImageLocation;
+            //    workBook.SaveAs(fileName);
+            //    workBook.Close();
+
+            //    MessageBox.Show("Файл " + Path.GetFileName(fileName) + " записан успешно!");
+            //}
+            //catch (Exception ex)
+            //{
+            //    MessageBox.Show("Ошибка: " + ex.ToString());
+            //}
+
+            MessageBox.Show("Файл " + Path.GetFileName(fileName) + " записан успешно!");
         }
 
         private void списокToolStripMenuItem_Click(object sender, EventArgs e)
