--- Объявление/ADList.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace App_project
-{
-    public class ADList
-    {
-        List<AD> adList;
-
-        //конструктор
-        public ADList()
-        {
-        }
-
-        //Просмотр списка карточек ДЖ
-        public List<AD> LoadADList(IFilter filter, ISort sort)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
